refactor-fix: [DEV-9730] Chart Refactor Fixes (#1874)
* rename cdc chart wrapper

* added additional refactor changes

* Chart Refactor Fixes
