+++ packages/chart/src/data/initial-state.js
@@ -306,5 +306,6 @@ export default {
       textAfter: ')'
     },
     data: []
-  }
+  },
+  runtime: { yAxis: {}, seriesKeys: [] }
 }
