+++ packages/chart/src/store/chart.reducer.ts
@@ -3,7 +3,7 @@ import defaults from '../data/initial-state'
 import { type ViewportSize } from '../types/ChartConfig'
 
 export const initialState = {
-  isLoading: true,
+  isLoading: false,
   config: defaults,
   stateData: [],
   colorScale: null,
